Fix yr_rules_scan_proc (#785)
* Add simple test for yr_rules_scan_proc

* Fix process scanning on Linux/i386

- Configure large file support (-D_FILE_OFFSET_BITS=64). Enough for
  scanning 32bit processes.

- Change data types so that they can hold large offsets. Required for
  scanning 64bit processes such as the shell that is spawned from
  test-rules on the Travis CI host.

Closes #784

* Do not fail process scan test if attaching failed.

Apparently, MacOSX rules for task_for_pid() have been changed over
time such that a user is not necessarily allowed to read memory from
his own processes.

+++ libyara/include/yara/types.h
@@ -408,8 +408,8 @@ typedef struct _YR_MEMORY_BLOCK* (*YR_MEMORY_BLOCK_ITERATOR_FUNC)(
 
 typedef struct _YR_MEMORY_BLOCK
 {
-  size_t size;
-  size_t base;
+  uint64_t size;
+  uint64_t base;
 
   void* context;
